fix(activities): the mailbox read answers one snapshot and every connector shape
Three findings from review.

The provenance fallback picked a fixed segment, and the segment count varies by
connector: a mailbox stamps connector:gmail:<seat> while an extension stamps
connector:ext:<unit>:<seat>. Counting from the left read a unit name on one and
nothing on the other. The seat is the LAST segment, so the match is on the whole
stamp -- and a stamp with anything after the uuid now names nobody rather than
attributing the row to whoever it mentions.

The read is composed from two statements and the first one is the authorization,
so it runs at REPEATABLE READ. Under READ COMMITTED a capture committing between
them -- a new import row, then the audience recompute behind it -- would let the
second statement answer a seat for a state the audience check never approved.

The disclosure got the Storybook story the design system asks of a new visible
state, and its tests moved into their own file rather than growing a suite that
was already far past the size ceiling.
